appli/requirements.txt
fastapi==0.110.0
starlette==0.36.3
jinja2==3.1.3
uvicorn[standard]
pymongo
python-dotenv
authlib
google-genai
itsdangerous
httpx